Stop ignoring pointer-sign warnings
Ignoring this warnings was added for getgrouplist(3) API differences on
MacOS vs other systems. On MacOS the third argument, the list of GIDs,
is a pointer of `int` while elsewhere it is a pointer to `gid_t`. The
MacOS API is definitely weird since e.g. setgroups(2) takes a pointer
to `gid_t` (even on MacOS).

Use a cast for MacOS only.

See e7f5274b ("Drop unnecessary cast") and 6e42c184 ("return the cast
(triggers warning in Xcode without the cast)") for previous attempts on
this warning.