util: Flush and sync id file before closing it
This way it should be written to disk (flash) instantaneously.

Note: on embedded devices using UBIFS every now and then we had the
case where the device was turned off shortly after monit generated a new
id, but before the kernel flushed it to the flash.  The usual outcome of
this is a file of zero byte size, like this:

    root@mxg-1 ~ $ ls -l /mnt/data/var/monit/id
    -rw-rw-r--    1 root     root             0 Jan  1 00:30 /mnt/data/var/monit/id

That 0 byte sized file lead to frequent errors in log later like this:

    [2012-01-01T03:00:01+0000] error    : Error reading id from file '/mnt/data/var/monit/id'

The ubifs maintainers recommend to flush and sync files with important
content immediately.  fflush() pushes cached stuff from libc to kernel
here, and fsync() from kernel buffers to disk.

Link: http://www.linux-mtd.infradead.org/faq/ubifs.html#L_empty_file
Signed-off-by: Alexander Dahl <ada@thorsis.com>